requirements.txt
Flask==3.0.3
PyJWT==2.9.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
python_bcrypt==0.3.2